Inital creation of the script
This is a vash script that renames files on a mounted GDrive in Linux from 'copy of' to the original name. 
You would need to do this when you make copies of files that you are moving from a different Google SSO GDrive account to another. 

In my case, I am moving all my GDrive files from my old free Google account to my paid 'Workspaces' account, and this was the easiest way to migrate the files over, since Google doesnt let you do this function.